html/editing/dnd/drop/events-input-same-element-manual.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Selection drag and drop: events for &lt;input> (same element)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body data-expected-events="b:drop,b:beforeinput,b:input,b:beforeinput,b:textInput,b:input">
<input id=b data-select="0,7" value="Drag me ...to here:">
<script src="/uievents/textInput/support/common.js"></script>
<script src="support/events.js"></script>
